Move queries and types to separate folder
