+++ src/lib/sanity.queries/logo/types.ts
@@ -1,0 +1,8 @@
+import type { ImageAsset } from '@sanity/types'
+
+export interface LogoType {
+  _type: 'logo'
+  logoImage: ImageAsset
+  caption?: string
+  href?: string
+}
